[bugfix] Potential overflow on controller buffer
elt_size and num_elts were stored as uint8_t, so their product could
overflow

+++ firmware/xbot2014/xbot2014/pipe.h
@@ -5,20 +5,20 @@
 
 typedef struct
 {
-    uint8_t write_to;
-    uint8_t read_from;
-    uint8_t elt_size;
-    uint8_t num_elts;
+    uint16_t write_to;
+    uint16_t read_from;
+    uint16_t elt_size;
+    uint16_t num_elts;
     void *data;
 } pipe_t;
 
-void pipe_init(pipe_t *self, void *data, uint8_t elt_size, uint8_t num_elts);   // Set up pipe
+void pipe_init(pipe_t *self, void *data, uint16_t elt_size, uint16_t num_elts); // Set up pipe
 void pipe_flush(pipe_t *self);                                                  // Clear data in pipe
-uint8_t pipe_write(pipe_t *self, const void *data); // Returns num bytes written, or 0 if full
-uint8_t pipe_read(pipe_t *self, void *data);        // Returns num bytes read, or 0 if empty
-uint8_t pipe_peek(pipe_t *self, void *data);        // Gets data without removing it from pipe
-uint8_t pipe_isEmpty(pipe_t *self);                 // Checks if pipe contains data
-uint8_t pipe_isFull(pipe_t *self);                  // Checks if pipe has room to write
+uint16_t pipe_write(pipe_t *self, const void *data); // Returns num bytes written, or 0 if full
+uint16_t pipe_read(pipe_t *self, void *data);        // Returns num bytes read, or 0 if empty
+uint16_t pipe_peek(pipe_t *self, void *data);        // Gets data without removing it from pipe
+uint8_t pipe_isEmpty(pipe_t *self);                  // Checks if pipe contains data
+uint8_t pipe_isFull(pipe_t *self);                   // Checks if pipe has room to write
 
 
 #endif
